fix: collecteurs fragiles — Nitter health tracking + arXiv delta normalisé (#206)
- nitter_rss: ajout d'une 4ème instance (nitter.cz), timeout 10s → 6s,
  rotation à partir du dernier index OK (_last_ok_idx), émission de
  source_health={"nitter_rss": True/False} via le merge dict de StateUpdater
- arxiv: émission d'un delta normalisé (PE) plutôt que d'un count brut
  via rolling average des 7 derniers appels (_count_history), valeur centrée
  sur 0 active même en dehors des fenêtres de soumission
- state: arxiv_papers_today int → float (compatible delta normalisé)
- tests: 10 nouveaux cas couvrant la rotation d'instances, la persistance
  de _last_ok_idx, le calcul du delta, et le bornage de l'historique

Closes #204

+++ core/state.py
@@ -33,7 +33,7 @@ class GlobalState(BaseModel):
     wikipedia_views_ai: float = 0.0
     newsapi_volume: float = 0.0
     newsapi_sentiment: float = 0.0
-    arxiv_papers_today: int = 0
+    arxiv_papers_today: float = 0.0
     github_ai_stars: float = 0.0
 
     gdelt_global_tone: float = 0.0
